Allow fields to be programaticaly re-validated.
Given two fields Field A and Field B, a validation rule such that Field B must
be greater than the value of Field A.  If both Field A and Field B are filled
and valid.  If the value of Field A is changed; Field B must be re-validated.

The following code:

  $("#field-b").parsley("validate");

does not cause Field B to be validated.

Removing these lines will fix this problem.