Fix display:block being added to divs
Parsley code adds the css display property with default value for the html element when using .show(). When using jQuery the default value for a div element will be block, ignoring the css markup of such element. This can break the markup and can be fixed by using .css('display', '') instead of .show()